Only copy boot.efi if present

+++ resources/install.py
@@ -175,7 +175,8 @@ Please build OpenCore first!"""
             print("- Copying OpenCore onto EFI partition")
             shutil.copytree(self.constants.opencore_release_folder / Path("EFI/OC"), mount_path / Path("EFI/OC"))
             shutil.copytree(self.constants.opencore_release_folder / Path("System"), mount_path / Path("System"))
-            shutil.copy(self.constants.opencore_release_folder / Path("boot.efi"), mount_path / Path("boot.efi"))
+            if Path(self.constants.opencore_release_folder / Path("boot.efi")).exists():
+                shutil.copy(self.constants.opencore_release_folder / Path("boot.efi"), mount_path / Path("boot.efi"))
             if self.constants.boot_efi is True:
                 print("- Converting Bootstrap to BOOTx64.efi")
                 if (mount_path / Path("EFI/BOOT")).exists():
